requirements.txt
streamlit
pandas
openpyxl
plotly
google-api-python-client
google-auth
google-auth-oauthlib
anthropic
streamlit-authenticator
